# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "SchoolBridge API",
    "description" : "WhatsApp-first school-parent communication platform. All 2xx responses (except 204) are wrapped: { \"data\": T, \"meta\": PageMeta | null }. Errors use RFC 7807 ProblemDetail. Tenancy is carried by the bearer token — no tenant header is required.",
    "contact" : {
      "name" : "SchoolBridge Engineering"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Default"
  } ],
  "security" : [ {
    "BearerJWT" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/assistant/knowledge" : {
      "get" : {
        "tags" : [ "document-admin-controller" ],
        "operationId" : "list_6",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/KnowledgeDocumentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "document-admin-controller" ],
        "operationId" : "ingest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IngestDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/KnowledgeDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assistant/knowledge/{id}" : {
      "delete" : {
        "tags" : [ "document-admin-controller" ],
        "operationId" : "delete_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "IngestDocumentRequest" : {
        "required" : [ "content", "title", "type" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "KB", "FAQ", "TOOL_DOC", "GUIDE", "BUSINESS_RULE", "UPLOAD" ]
          },
          "title" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "lang" : {
            "maxLength" : 8,
            "minLength" : 0,
            "type" : "string"
          },
          "content" : {
            "maxLength" : 200000,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "KnowledgeDocumentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "KB", "FAQ", "TOOL_DOC", "GUIDE", "BUSINESS_RULE", "UPLOAD" ]
          },
          "title" : {
            "type" : "string"
          },
          "lang" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "INDEXED", "FAILED" ]
          },
          "chunkCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ParentToken" : {
        "type" : "http",
        "description" : "Opaque 24-hour sliding-TTL token issued by POST /api/v1/parents/auth/verify-otp.",
        "in" : "header",
        "scheme" : "bearer"
      },
      "BearerJWT" : {
        "type" : "http",
        "description" : "RS256 JWT issued by POST /api/v1/auth/login (staff/admin).",
        "in" : "header",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}